Added a bottom sheet for the logout so if the user wanted to change his mind and not to logout he presses on the cancel button
