+++ app/src/main/java/com/example/eshfeenygraduationproject/eshfeeny/more/LogoutBottomSheetFragment.kt
@@ -1,15 +1,17 @@
 package com.example.eshfeenygraduationproject.eshfeeny.more
 
+import android.content.Intent
 import android.os.Bundle
-import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import com.example.eshfeenygraduationproject.R
+import com.example.eshfeenygraduationproject.authentication.AuthenticationActivity
 import com.example.eshfeenygraduationproject.databinding.FragmentLogoutBottomSheetBinding
+import com.example.eshfeenygraduationproject.eshfeeny.publicViewModel.viewModel.UserViewModel
+import com.google.android.material.bottomsheet.BottomSheetDialogFragment
 
 
-class LogoutBottomSheetFragment : Fragment() {
+class LogoutBottomSheetFragment(val viewModel: UserViewModel) : BottomSheetDialogFragment() {
 
     private var binding: FragmentLogoutBottomSheetBinding? = null
 
@@ -19,6 +21,22 @@ class LogoutBottomSheetFragment : Fragment() {
     ): View? {
 
         binding = FragmentLogoutBottomSheetBinding.inflate(inflater)
+
+        binding?.apply {
+            this.signoutButton.setOnClickListener {
+                viewModel.deleteUserFromDatabase()
+                val intent = Intent(
+                    activity,
+                    AuthenticationActivity::class.java
+                )
+                intent.flags = Intent.FLAG_ACTIVITY_CLEAR_TASK or Intent.FLAG_ACTIVITY_NEW_TASK
+                startActivity(intent)
+            }
+
+            this.cancelButton.setOnClickListener {
+                dismiss()
+            }
+        }
 
         // Inflate the layout for this fragment
         return binding?.root
